requirements.txt
dj-database-url==0.5.0
Django==2.1.4
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
python-decouple==3.1
pytz==2018.7
whitenoise==4.1.2
